[feat] create 페이지 컴포넌트 분리 완료  #49
